+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### ✨ Features
+
+- Improved the setup of the plugin. Now the plugin will prompt you for an API token on launch and you can set the token through the settings.
+
 ## [1.5.0] - 2020-10-31
 
 ### ✨ Features